Adding in the basics for upstream projects
Bumped to aweplug version 1.0.0.a9
Added a new extension to create the basics for upstream projects.

It strikes me as this being something that could be a little easier, and
probably wouldn't need the synthetic handler. The trade off then becomes
looping through every page in the site to find the right page(s) to add
things to using the searchisko helper. I'm not sure the best way to go
about doing this. Should probably talk it over with someone.