DEVELOPER-249 Created a new Slim partial
This is done to allow different pages to use different partials.
One set of pages requires the partial to be embedded in JavaScript
and templated at runtime. The other set is generated at site build
time.

Rewired event handlers to ensure that dynamically inserted
Socialte buttons will be displayed correctly on hover/click.

Removed the Socialite sprite since it is displayed only initially.