Fixes for DEVELOPER-487
Contains aweplug fixes
Adds all technologies in our quickstarts

I took some things from what Lukas had given us, but ultimately we still
need the query. We may be able to figure out exactly when we need it and
when we don't but there are a lot of edge cases I think. Also, without
the query we don't get the score, which we may want later.

As a bonus, using more than one topic now gives us the "AND" behavior
people are expecting :)