Fixes XSS on product get started pages DEVELOPER-966
Squashing... for no good reason.

I added the jquery encoder plugin and am doing encodings for the
different areas we're using the query string.

Lower casing javascripts

Found this problem locally where it would 404 on lower cased scripts,
but it isn't a problem in production or staging (I assume because we concat
everything). When I lower cased the names in some of the files in my
last commit it broke the build on staging. This should fix it and
everything should be happy again in all environments.

Downloads should work once again.

I reworked how we build everything up a bit, also checking to make sure
we're actually using a valid Red Hat URL, and the file name param and
url match.

It may not be 100% fool proof, but it's much better than what we had
before.