Fixes DEVELOPER-1121 (buzz tags for products)
It's a bit of a round about way of doing it, but this works. We're
essentially storing JSON in the data attribute now. We have to massage
it a bit to make the quotes work correctly (switching from double to
single then back again). `buzz_tags` in the yaml file are now stored in
arrays.

I also fixed some jshint issues in the buzz.js file.