Fixes DEVELOPER-1109 and fixes DEVELOPER-1016
DEVELOPER-1109 deals with the contribution guide, which should be
rendering correctly now, We're not touching absolute links from
kramdown.

DEVELOPER-1016 is about quickstarts in JDG (or any other place) that
links directly to the README.md files using relative links. Now we're
simply doing a gsub to replace "README.md" with "index.html". The rest
of the link should work correctly.

Adding back in the CONTRIBUTING fix