Fixes DEVELOPER-1145 (product download right bar alpha)
Downloads on the right product bar now support alphas, or RCs, or
whatever. We maintained the alpha and beta overrides in the yaml file.

We're now simply looking for releases that are newer than the current release. Those are grouped together, sorted by date, and the first is taken (the newest one) and used as the newest version for the download.

We also look to see if the version is an alpha or beta by searching for
'alpha' and 'beta' in the version string and flag the download
accordingly.