Fixes DEVELOPER-1192 indentation on get-started
Probably the best we can do is this small style hack. What was happening
is that you get on ol:li inside of another ol:li section which really
messes up the formatting. A correct fix would be to end one ol and start
another, but I don't see a way of doing this easily without refactoring
the layout extensively. I did try some other solutions, but nothing
worked the way I expected it to.