Fixes RHD-76: Dockerizing the build
I have created a docker container for the awestruct build.

The docker container needs to have a volume for the site source.

We're building up a new environment for everything to happen in complete with rvm and bundler.

Creating a new control script

Switched the control script over to ruby, which lets us do a few more things easier.
The control script also makes use of the vault in ../_config/secrets.yaml.gpg (which has also been updated) to setup the correct environment for the awestruct build using docker-compose.

 It is highly recommended that building the site to be used with drupal be done using this script as it has all the necessary environment variables already setup.

 I recommend at a future date to create a new awestruct profile that we can use to check if we're building for drupal, if we are then we ensure we do some other processing (like checking to see if drupal is alive, pushing to searchisko, setting the correct urls for the services, etc).

Using the correct service URLs for docker-compose

I recommend creating and using a different profile for docker use in the
future.

Updating drupal a bit

I've updated to use a better feature cut of our awestruct endpoints, one
which doesn't use pathauto which is very time consuming.

I also combined some of the extension enabling and downloading.

It would be much better if we could get all of this into the docker
container at build time instead of runtime.

Taking out logging on mysql

Logging is an extremely performance killing operation in mysql. I'm
removing it to help increase perf a little bit when we push to drupal.

Right now the whole build, assuming you've already built the containers,
takes about 10 minutes on my box. That's building the site and pushing
it into drupal. I have a feeling there's something else going on inside
drupal which is costing us a lot of time as I don't recall it being this
slow before.

Adding a couple new extensions for drupal

I'm doing the drupal push during a transform now instead of an extension
which happens before the transform. This could easily become an
after_generation extension in the next version of awestruct.

I'm also testing to see if drupal is alive before proceeding forward in
doing the push, this way simply using 'control.rb -r' will do the whole
build for us.

Doing checks in the control script for liveliness

The control script now checks to see if containers are active and ready
for data before we move on to the other parts of the build. I need
someone to test this on OS X to make sure things are working well for
boot2docker or other ways of running the docker container.

Removing the checking from awestruct

Docker script and containers are working

Finally everything seems to be working!
I switched to head Awestruct, will switch to the released version as
soon as I get some bugs ironed out.

I had to change the drupal volume location for some reason, don't know
why it wasn't working, but whatever.

I'm no longer using rvm as that seemed to be causing some issues, so I
dropped it. If we do change ruby versions we'll need to update the
containers.

Sending the pages to drupal adds an extra two minutes to the build. I
don't like it for development, but maybe it isn't that big of a deal,
not sure yet.

Everything appears to be working for dev

I've not yet tested this for use with drupal.

Fixing up some issues after the code review

Updating README, fixing issues with awestruct start

Fixing issue with building & copying Gemfile.lock

Adding info about using current git version

Clean-up from the rebase

Don't need these files