Fixes RHD-186 (learn links going to www.jboss.org)
I added another filter that is used for sys_url_view which should
rewrite the host (any also the PR url) to use site.base_url. I believe
this should be correct, I'll need to double check once we have a
build though, as I'm not sure if it'll be correct for staging builds.

Small issue fixes

Adding in '/' at the end of the replace.
Fixing double '/'

Remove the '//' fix as that breaks http(s):// :(