Fixes RHD-156 - generation of product index page
Most of these are small fixes and adding data.

There's a new `_config/categories.yml` file which contains the names,
description (optional), and display order of each category. It also contains the list of
products (and incidentally display order) which belong to that category.

I removed the equal height hacks for the product index page as they no
longer are needed.

Each product now has an `index.desc` field which is used as the
description for that product on the index page. If there is no
`index.desc` field then the `description` field is used. This data is
passed to Asciidoctor and the rendered output is used in the webpage.

Solutions now also have a `related_products` field which is an array of
products this solution applies to. The current data is simply test data
and should be fixed.

Fixing links, adding commas, fixing buttons

* The solution links should now link up correctly.
* See It In Action are now comma separated.
* Download buttons are there for all products that have a
  default_download_artifact and latest_version field. Products
  which do not meet that criteria have no download button

Fixing up links

All links should now be absolute links instead of relative links