Fixing BOM replacement URL
Previously the code to link to a bom which replaced an old bom was
broken and the links did not exist. This fixes that problem, though I've
just thought of another solution: Why send boms to the DCP which have
replaced product boms anyway? Yes, we'd need to erase the current set of
BOMs in the DCP, but if we're going to have a link to a replacement bom,
why even include the old bom in the first place?

debug: trying to determine url building issue