Fixes DEVELOPER-1814 (awestruct cdn)
Some of the work happened in aweplug, so you'll need to see the changes
there.

We're using drupal as the CDN now. All the images are being added to
drupal on drupal startup via a combination of a drupal hook_install
implementation and also a docker mount from the images directory, you
can see those changes in a different pull request.

We needed to add the cdn info to the drupal awestruct profile, we also
needed to add the cdn_prefix to the environment when we start up drupal
via the control script.