Fixes DEVELOPER-2505 (angular dcp not working)
I took out the wrapping of everything in a function as that was causing
some scope issues in the javascript. Because of that I had to create $
for our jquery stuff. I'm also no longer uglifying the concated
javascript files because it was causing problems with the angular module
as well. We can look at what's really happening with that at some other
point.