[DRUPAL] Topics and misc (#1004)
* force options to be as per -u --run-the-stack

* adding a build task to drupal-nightly

* Using Open3 to show errors for gulp build

* Having npm tell us where gulp is

* drupal_host_ip || hostnam  && preview -> gen

In the drupal profile if there's no ENV['DRUPAL_HOST_IP'] use ENV['HOSTNAME'] instead of just docker as this will mess things up on the developer's box.

Also for drupal we should be doing a gen instead of a preview.

* Using ENV if available for drupal ip

* Removing searchiskoconfigure for drupal-nightly

Also small spelling fix

* Adding topics and container news

Also dabbling a bit in creating a profile

* Removing docker services from previous run

On :kill_all we're going to be removing the services as well to try and a completely create a clean environment

* Removing containers when doing a kill

We no longer are using searchiskoconfigure so I fixed that test

* Had to rearrange the order of things for a view

* DrupalConsole finally has ssl :)

* WIP for both jenkins and local dcp location

* Moving things around

* Accidentally added these files when I moved them

* Supporting new layout structure for css/js

* Supporting restructure

Also trying to use the host name for postgres

The idea here is to only ship docker images instead of having to rely on
the source to be there as well. During development we're going to
override the installation of the themes and modules.

* Switching to docker-compose version 2 format

Adding changes to support version 2 changes for networking.

* I accidentally took out the images, oops

* Adding some debug info to see what's happening

* Adding verbose output for docker

* Revert "Adding verbose output for docker"

This reverts commit b84b2c031cf73654714fb39d736a8b364148fd09.

* Trying a different verbose

* Revert "Trying a different verbose"

This reverts commit c713dfd26a485fce4f4e67ab4d72a96aee35a899.

* Hacking around ENV settings for jenkins

* Removing the networking from docker-compose

This effectively goes back to what it was before version 2 of the
docker-compose file.

* Adding more to an install profile

This isn't exactly working yet, but I want to at least get it into
version control.

* Adding composer files

This is helpful for anyone doing work on drupal so they can create the
same environment.

* Drupal install profile into docker

It's commented out currently as some things aren't working yet.

* searchisko wasn't working in docker locally

Had to make a small fix for local environments

* Ignoring composer files

* Fixes DEVELOPER-2796 (missing JS files)

Adding the missing JavaScript files from the main site.

* Whitespace changes (tab 4 vs tab 2) && main page

The main change here is that we are now setting the main landing page
for the site.

* Attempting drupal page layout & display suite

* Topics quickedit works!

I no longer need the layout plugin.
Turns out I was just missing the title_suffix / title_prefix for
quickediting to work

* Removing items no longer needed

We don't need the view any more nor the block. I'm also no longer using
the layouts plugin.

* Not using the profile either

* Enabling topic fields for viewing

* Move drupalConsole install into run

* Fixes to work on linux && mac

Found some issues with the enviroment variables and host names between
Mac and Linux. I'm trying to patch these up so they work in both places.

* Commenting out unnecessary function binding

The function bind in ytembed.js was breaking Drupal quickedit, this has
been fixed by commenting it out.

* Fixes for mac

Trying to get the host name correct

* Small theme adjustments for topics

* Style tweaks for topics in drupal

There are still some things that need to happen, but I can't do those
without affecting the non drupal site

* Fixed the logic from nil to empty

Fixed site.yml to work with mac

* Fixed js error for when eventDate doesn't exist.

Updates for JS

* Renames from today's call

Also trying to create a sane layout for theme files.
Mike had some renames that he'd like to see for the topics, I've renamed
things to be more consistent with what is on the UI.

* Removing the merge conflict header

* File renames in missed in the last commit

* Changes for topics

Better layout, which more closely matches current live site, also the
four white cards are now called out.

* Moving the header region to a tab

Hopefully this makes it so things don't take up as much space for people
logged in.

* Adding JS for tab

* Spans to divs for layout

* Fix for the build

Indentation was wrong.

* Take two on build fix for yaml

* Take three I think :(

* Updating awestruct for logging

* Trying to get debugging

* Rakefile should use erb for site.yml

* Forgot to require erb

* ERB issue