Fixes DEVELOPER-2947 (fuse connector page)
Please see https://issues.jboss.org/browse/DEVELOPER-2947 for details.

In Drupal it was 'fixing' bad HTML. Well, turns out it thought the HTML
in the script tags was bad (probably because of our template marker
choice), so it did it's best to fix it. Turns out it actually broke it.
I've created a new html text format that only administrators can use
which doesn't do this correction. I named it 'as-is HTML' so Drupal
doesn't try to fix anything.

In order to test this you'll have to start up Drupal:

```
cd _docker
bundle exec ./control.rb -u --run-the-stack
```

Once that finishes you'll need to visit the fuse connectors page to
verify things are working correctly.