docker env determine_docker_host fix
control.rb::determine_docker_host_for_container_ports was trying to
get the IP address of the system hostname, which might defined on
developer's machines, before checking whether hostname `docker`
was defined.

The expectation is developer's machines with an RHDev docker
environment will have docker defined as a hostname with an
ip address for localhost.

This workaround will still fail if the hostname docker isn't
defined and the current hostname doesn't resolve.