Rewriting URIs to production location in sitemap
All URIs in the saved sitemap are now being rewritten to point to the
production location.

Also added a test to verify this behavior.

Added some gems to the Gemfile to aid in running these tests from the
command line.

Maybe I'm the only one having this problem, but I've found that using
mocha in a debug session causes the debugger to silently fail (unless
you have verbose logging on).