Fixes DEVELOPER-3292 DEVELOPER-3291 DEVELOPER-3293
Turns out all of these had some little bit that were related, so I'm
fixing them all in one commit.

See

* https://issues.jboss.org/browse/DEVELOPER-3292
* https://issues.jboss.org/browse/DEVELOPER-3291
* https://issues.jboss.org/browse/DEVELOPER-3293

for more detail about each issue.

A good portion of these had to do with drupal fixing up hrefs in anchor
tags. That's all well and good for most things, but for pages that use
Angular, we need it to not do that, hence bringing back the as_is_html
filter. This filter is only used for pages that have the drupal_filter
set to `as_is_html`, otherwise it uses `full_html`. A few files were
updated to include this new variable in the front matter.

There were some added client side templates in the current site which
were not added for Drupal, so I needed to recreate those for Drupal.
There was also a new variable, the `broker` variable, which needed to be
added. That change happend here for dev and for staging and prod I made
those changes on the respective systems.

Lastly, with regard to the search page, some changes in the base chrome
were not implemented into the Drupal side, so I corrected that. A small
change to the javascript search needed to change as Drupal will not use
urls like 'http://.../search/?q=foo'. I modified the regex to not
include that last slash. This should be tested to make sure it still
works for the existing system. My initial testing says yes it does, but
it needs to be verified.

StackOverflow integration required some JavaScript fixes / additions for
Drupal and the client side template. On my testing I don't have any SO
issues in Searchisko (not sure why, I pulled a new dump) so I haven't
seen it work perfectly, but I know that if there are no issues in SO it
displays correctly :) Same with the search, for whatever reason my
latest searchisko dump isn't returning search results.

Feel free to go through this with a fine toothed comb to make sure it
all works correctly.