Fixes DEVELOPER-3397 (topic text formats)
Please see https://issues.jboss.org/browse/DEVELOPER-3397 for more
information.

This changes the formats for the `blue card description` field and the
`topic sub title` field. This is a configuration change, and as such
will require a rebuild of the current drupal instances (there may be a
better way to do this, but I'm not sure how, @drupalbee do you know?).