Fixes DEVELOPER-3441
See https://issues.jboss.org/browse/DEVELOPER-3441 for more information.

You may not see this is staging / prod until we rebuild the database.
This is actually a fix for a different issue I discovered, the blue card
description was being rendered as plain text, which isn't right. I had
to make a few other changes to it as well.