Improving the push to drupal
Seems like we're still having problems with the Drupal push. This uses
the node url (and prints it) and also the previous data from the
existing node (according to the docs this is the proper way to do it).
I'm hoping this helps cut down on new url alias entries.