[DEVELOEPER-3310] Product Pages Drupalized (#1617)
* initial changes

* adding product pages

* removing unwanted files

* fixing issues

* Reverting to master version

The config import was disabled, had to revert.

* Updating some dependencies

Updating some of the added dependencies to more current versions and
fixing a whitespace issue.

* Installing modules

* Allowing product pages to be created/modified

* Getting the buzz page to work

I had to add a field to the paragraph type and everything that goes
along with it, but we've got it.

* Adding changes for Docs and Api files integration

* Syncing active config

These always seem to not import on my box for whatever reason, so I'm
adding them to the active config.

* Removing deprecation notices

* Update Drupal and Modules

Drupal => 8.2.1 (because reason)

Most modules are built for the current release, and unless you're one a
LTS version of Drupal as soon as the new version comes out you have to
update if you want things to continue to work with new versions of
modules. Very frustrating IMO.

* Installing modules

* In dev, always turn twig debug on

* Product page basic functionality

I'm now where Jim was but using more standard Drupal coding standards
and practices. As a bonus, some of the product pages can be quick
edited.

* Better way to handle titles for left nav

The inline template type is a better way of doing this and staying safe.

* Updating to Drupal 8.2.2 && Product Pages

Some product pages are working, still left are:
* Help
* Getting Started
* Downloads
* Community
* Docs

* Take out the debugging

* Adding the community product page

* Adding the display of main content for connectors

* Implemented Getting Started pages

* Product Download pages working

* Forgot some headers

* Putting html.html.twig back

* Added Overview, Resources and Updates

* Help page done

I made the stack overflow integration a bit less brittle too. We're
pulling the product code from the page instead of the address bar now.

* Found a better way to do angular code in twig

Swapped out the verbatim stuff to have twig print '{{' and '}}'.

* Docs fields

* Reworking some things

Reworked overview page
Docs and APIs working
Adding some modules to support overview rework

* Updates for clean entity upgrade

* Fixing up problems Paul has spotted.

* Fixing deep linking in tabs

* Adding metatags and sitemap for product pages

Product sub pages now how metatags and they are in the sitemap.xml as
well now.

* Updating permissions for products

* Disable caching

This allows the links to work correctly on the left hand side, and
allows it all to work correctly.

* Fixes for bugs Mike found

* Fixing display of links on the side

* Fixes for bugs Mike found