Fixes DEVELOPER-3731 and DEVELOPER-3736
Decided to do DEVELOPER-3736 while I was in there and this incorporates
a lot of the work needed for it anyway.

SO section now takes up the full allotted width if there are no forums.
Also if it is full width we're displaying two columns and eight
questions instead of one column and four questions.

If there is no forum link and no SO to display we're no longer
displaying "Community Q&A".