DEVELOPER-3948 setting up Drupal Unit Tests
We probably need to break up
https://issues.jboss.org/browse/DEVELOPER-3948 but here's a stab at
getting the infrastructure in place.

Unit tests for the rhd_common module can now be run via our control
script. We probably need to make some changes to the jobs in order to
get these to run as part our builds.

Drupal Unit Tests can be run using either `-d` or `--drupal-unit-test`.
This is probably the easiest way to run tests as running them locally
involves some knowledge of how to run it all.

Rob and I agreed that we should use the current infrastructure and
practices for functional and end-to-end testing.