Attempting to fix rhamt revert issues
I have already created a new version of the product page ready to go.
This change "deletes" the old version of the product page and publishes
the version I created.
It also fixes an issue where we may pull deleted product pages and
display them, which wouldn't be good at all.