DEVELOPER-4022 - Throw exceptions if empty/null
Also a bit of a reflow.... sorry.

If we don't have any revisions for a product or don't have an
active_paragraph/sub page throw a NotFoundHttpException. This will stop
the PHP errors and send a 404 back to the user, which is probably what
we want to have happen anyway.