Adding completed login/register tests.
Fix for dependancy hell. octokit was resolved to 4.7.0, which depends on
                         12:39:30     sawyer was resolved to 0.8.1, which depends on
                         12:39:30       addressable was resolved to 2.5.2, which depends on
                         12:39:30         public_suffix

Fix for dependancy hell #2 . octokit was resolved to 4.7.0, which depends on
                         12:39:30     sawyer was resolved to 0.8.1, which depends on
                         12:39:30       addressable was resolved to 2.5.2, which depends on
                         12:39:30         public_suffix

Fix for dependency hell #3 and cucumber tags was not configured correctly within the node jenkins runner.

Fix for dependency hell #4 Installing ruby 2.1 and specific version of octokit.

Fixed unit-tests

Moved Before hook as it is not supported by wdio-cucumber.

Tidy up and fix for flaky login test.

Fix for siteNav is not defined error in previous test run.

Fix for ERROR: Cannot find module '../../support/pages/home.page'

it works on my machine ....