Moving title up higher for monitoring checks
Our monitoring solutions need to verify the correct page is loaded. They
would like something towards the top that is guaranteed to be there. The
title seemed to be a good option, but it needs to be within the first 5k
of the response. I'm moving it up higher so it's guaranteed to be in
that amount.