DEVELOPER-5255: Image Style URLs
In this commit, in the 5 changed template files, we are passing just the
plain-text string to the <img> src or a <href> attribute(s). Previously,
many of these attributes were having entire Node objects passed to them,
which resulted in instances where HTML comments, for example, were being
injected into the src or href attributes. In addition to the 4 entity
type + display modes detailed in DEVELOPER-5255:

* Node, Events, Tile
* Node, Product, Featured Tile
* Node, Video Resource, Teaser
* Node, Video Resource, Tile

I also discovered a similar issue with the Assembly, Static Item entity
type + display mode, so I resolved that in this commit as well.