DEVELOPER-5432: Add languages as visual styles
I have added about 60 of the most common programming languages as Visual
styles to the Code Snippet assembly type so that an editor, in the event
that Highlight.js does not automatically detect the language, can go
back and select the language for their snippet. This will be applied as
a CSS class to the <code> tag which will tell Highlight.js which
language it is.