DEVELOPER-5432: Add programming language field
I have removed the list of programming languages from the visual styles
of the Code Snippet assembly type in favor of implementing them as their
own field. An editor can optionally select a language for the code
snippet using a drop down menu. The field only allows one value. If the
editor selects a programming language, the key of that value is passed
to the template via the preprocessor and added as a class upon the
<code> tag, as specified in the Highlight.js docs.