DEVELOPER-5450: Clean up custom theme directories
This commit removes the RHD theme: both uninstalls it from Drupal and
removes it from the directory tree. We are no longer using it, and this
will make our image smaller.

In order to accomplish this, I had to move some Twig templates, that
were being included, over to the RHDP theme and change some references.