DEVELOPER-5488: Update Core, Lightning, Paragraphs
This commit updates Drupal core to version 8.6.2, the latest version, it
updates the Lightning profile and dependencies to the latest version, it
updates the Paragraph module to version 1.5.0, the latest version which
includes a security release.

This commit also includes the config-export performed after executing
`drush updb` and `drush update:lightning` after updating all of the
dependencies with Composer.