Clean up rhdp.theme file
This commit generally tidies up the rhdp.theme file. It applies Drupal
coding standards to almost the entire file, with a few exceptions for
line width. It applies a 2-space indentation to the entire file. It
reorganizes a few portions of the preprocessors to improve readability.

I have also discovered an issue in rhdp_preprocess_views_view() that
will throw an undefined variable error upon any execution of this hook.
I have noted that in the appropriate place in the file. We should
resolve that within the PR containing this commit if possible.