DEVELOPER-5453: Default to prod
This changes the behavior I implemented in my last commit to always
default to the Keycloak prod script *unless* the
redhat_developers.keycloak.scriptUrl config object is set. If it is,
which we will set on the Stage environment, then we will use the value
of that config object to point to the stage Keycloak script URL.

I also removed some code from export_html_post_processor.rb in this
commit that was previously needed to rewrite the keycloak URL during the
export process. I do not believe this will be necessary now.