Downloads page controller, preprocess assembly
This properly makes it so that the ProductsDownloadsController renders
the Product node in the Products Download Page view mode, and does the
preprocessing of the assembl(ies) on that page within preprocessors.
These components should be loosely coupled in this way. We cannot assume
that the download list, etc. assembly types will exclusively be placed
on Product pages.