Remove Gulpfile.js in root and from package.json
This removes the Gulpfile.js from the root of the repo. This is a relic
of an old, deprecated front-end build process. The new front-end build
process uses npm within the rhd-frontend directory within the rhdp
theme. This also removes a reference to the Gulpfile in the package.json
file in the root of the repo.