Add AssemblyBuilds, Getting Started Page
This moves logic that I previously had in preprocessors, for the Product
Download List and Product Download Hero, into AssemblyBuilds. This seems
like a much better architecture, and the parent entity is already
available in the build() method.

I also created a Getting Started that very closely mimics the
implementation of the Downloads page I've created in this branch.

Since they are not necessary, I've removed the entity reference Product
field from all 3 assembly types. The parent entity (Product) is
available in the AssemblyBuild.