Adds visual styles for the Product heroes
This adds the following visual styles for the Product Download Hero and
the Product Try It Hero: light, dark, centered and slim.

There is remaining work required to actually implement the styles. This
commit only creates the visual styles on the assembly types, which will
provide classes on those components if an editor selects one or more of
the visual styles.