Preprocess Product Features assembly
This adds all of the configuration changes and preprocessing code
necessary to nicely provide all of the necessary variables and logic in
the templating layer to render the Product Features assemblies in HTML
tables. Currently, the table cell values are 1 if that Product has the
Feature or 0 if it does not. I think this will be a checkmark or empty
in the design though.