Revise routing strategy of Product pages
There was a conflict on the /products/[product-url-code]/download URL
pattern between the current implementation of the Product pages and this
new implementation. To resolve it, I had to add logic to both
controllers that handled each routing condition based upon whether or
not the 'Use New Product Page' boolean field on a given Product node
was, or was not, checked.

Furthermore, this commit provides more assembly types on the Downloads
Page Content field, as requested by Tiffany. These changes can be seen
in our running Google sheet.