Fixes <title> tag missing content
This fixes an issue that ultimately resulted in missing <title> tag
content for the /download and /getting-started routes. This sets the
dynamic page title, as we would expected in transitioning for our old
approach, so that Metatag can properly identify the content that needs
to go into the <title> tag.