Remove the package.json and package-lock files
This removes the package.json and package-lock.json files from the root
of the repo. As discussed in the PR, on Github, the package.json file is
no longer being used. If that is the case, then there is no point in
keeping the package-lock.json file.

This also implies we _should_ be able to remove the node_modules
directory from the root of the repo.