Featured Articles Assembly Type
THis creates a Featured Articles assembly type and allows this assembly
type to be referenced from the Content, Downloads Page and Getting
Started Page fields on the Product content type. The Featured Articles
assembly type allows an editor to reference existing Articles to create
a collection of articles that are pertinent to the page/node that the
Featured Articles assembly is placed upon.

I have also created a custom template for node--article--card.html.twig
and field--assembly--field-articles-featured-articles.html.twig to start
the templating/layout for the FE work. The remaining FE work includes
any styling required to meet the InVision mockups (attached to JIRA
ticket).