Remove Product Features assembly from Articles
This removes the Product Features assembly type as an option from the
Content entity reference revisions field on the Articles content type.
Gina has requested we removed this assembly type as an option on this
content type until we implement some changes to the styling of the
Product Features component that will allow it to be displayed more
functionality in situations with less width.