Resolving redeclaration of form alter
This resolves a redeclaration of hook_form_alter(). It was originally
invoked in the .module file, and I invoked it in my last commit in
.form_validations.inc. I have moved the logic that was contained within
the hook_form_alter() in the .module file into the .form_validations.inc
file since that logic pertained to form/form element validation(s).

Furthermore, I consolidated some switch case statements to make the
switch statement more readable, and I resolved a typo in a comment that
Jason pointed out to me.