Create Connectors Views Block, no AJAX calls
This commit creates a Block display for the Connectors REST View. This
block is programatically placed into the correct position in the
Connectors paragraph.

The sorting of the connectors is now done without any AJAX calls to
either DCP or Drupal, but instead by utilizing data attributes on the
li.connector elements.