Form Validation URL product form
This provides some validation of the URL product name field on the
Product form such that the URL product name field should be unique
across all Product nodes. If an editor tries to submit the form with a
value for URL product name that is already being used by an existing
Product, then an error will be thrown on the form submission.