Adds custom URL Product name validation to Product
This adds custom validation to the URL Product Name field on Product
node forms. When a Product node form is submitted, this custom
validation hook will try to retrieve any existing Product(s) that have
the same value URL Product Name field. If any such Product(s) exist,
then an error will be thrown on the URL Product Name form field and the
submission will be rejected. The editor will be presented with an error
message explaining that the value of the URL Product Name field on the
Product content type must be unique due to routing.