Product node validation false negative
This resolves a false negative scenario in the form validation of the
Product nodes in which Product nodes could no longer be saved because
the validation would recognize that a duplicate Product URL Name field
value existed (itself). To prevent this scenario, we are now verifying
that the Node ID of the match is not equal to the Node ID of the parent
node of the Pproduct node form.