Validation of SSO config before config override
This adds validation to the SSO config, coming from rhd.settings.yml,
prior to attempting to set a config override for the OpenID/Keycloak
config variables in Drupal. If at least one config value is unset or
empty that is required, a notice will be logged in Drupal and we will
not set any config overrides until all of the required configuration is
available.