Provide a config override warning on Keycloak
This provides a value in Keycloak config values (strings) of "Set with
config override: check rhd.settings.yml" so that admins, viewing the
config form in the UI, will know that these config values are set using
config overrides. Without this warning, it would appear that the config
is set improperly since the config override values do not display on the
config forms.