Fix syntax error in config override
I made a silly mistake / had a silly misunderstanding in the actual
slug/machine name of the config object vs. properties on the config
object, which resulted in the config object not being fetched, and
therefore not set, in my config overrides. I believe from my local
testing that this now should properly set the config override.

You can test this by using drush config-get with the
--include-overridden flag.