Update Core 8.6.5 and Lightnig 3.2.3
This updates Drupal core to version 8.6.5 and the Acquia Lightning
profile to 3.2.3, including many updates to dependencies of these two
major packages. This should be updated anyway, but I am specifically
working on this to address a 'failure to apply patch(es)' error that Rob
discovered and is blocking his MP work.