Change Keycloak button text, make button evident
This makes some changes to make the OpenID Connect Login button a little
more evident, but still not incredibly prominent to editors. This is
achieved by putting the button in a <details> element. Also, I created a
patch for the OpenID module to change the text of the OPenID Login
button, for only the Keycloak provider, to "Log in with your developer
account". Unfortunately, the render array for this block is being
generated in the actual Form class, as opposed to utilizing a Twig
template, so a patch to the Form class was required vs. overriding a
Twig template.